Moved ILogin interface in type/page folder
